remove TokenSwappingWithArch project; move files into Architecture
