+++ tket/src/Architecture/NeighboursFromArchitecture.cpp
@@ -29,9 +29,9 @@ const std::vector<size_t>& NeighboursFromArchitecture::operator()(
   const auto num_vertices = m_arch_mapping.number_of_vertices();
   // GCOVR_EXCL_START
   TKET_ASSERT_WITH_MESSAGE(
-      vertex < num_vertices, "get_neighbours: invalid vertex "
-                                 << vertex << " (only have " << num_vertices
-                                 << " vertices)");
+      vertex < num_vertices,
+      std::stringstream() << "get_neighbours: invalid vertex " << vertex
+                          << " (only have " << num_vertices << " vertices)");
   // GCOVR_EXCL_STOP
   auto& neighbours = m_cached_neighbours[vertex];
   if (!neighbours.empty()) {
@@ -53,9 +53,9 @@ const std::vector<size_t>& NeighboursFromArchitecture::operator()(
     // GCOVR_EXCL_START
     TKET_ASSERT_WITH_MESSAGE(
         neighbour_vertex != vertex,
-        "get_neighbours: vertex "
-            << vertex << " for node " << node.repr() << " has "
-            << neighbour_nodes.size()
+        std::stringstream()
+            << "get_neighbours: vertex " << vertex << " for node "
+            << node.repr() << " has " << neighbour_nodes.size()
             << " neighbours, and lists itself as a neighbour (loops not "
                "allowed)");
     // GCOVR_EXCL_STOP
