+++ tket/tests/TokenSwapping/TestUtils/FullTsaTesting.cpp
@@ -20,9 +20,9 @@
 #include "TokenSwapping/DistanceFunctions.hpp"
 #include "TokenSwapping/RiverFlowPathFinder.hpp"
 #include "TokenSwapping/VertexSwapResult.hpp"
-#include "TokenSwappingWithArch/ArchitectureMapping.hpp"
-#include "TokenSwappingWithArch/DistancesFromArchitecture.hpp"
-#include "TokenSwappingWithArch/NeighboursFromArchitecture.hpp"
+#include "Architecture/ArchitectureMapping.hpp"
+#include "Architecture/DistancesFromArchitecture.hpp"
+#include "Architecture/NeighboursFromArchitecture.hpp"
 
 using std::vector;
