+++ tket/tests/TokenSwapping/TestUtils/PartialTsaTesting.cpp
@@ -20,8 +20,8 @@
 #include "TokenSwapping/DistanceFunctions.hpp"
 #include "TokenSwapping/RiverFlowPathFinder.hpp"
 #include "TokenSwapping/VertexSwapResult.hpp"
-#include "TokenSwappingWithArch/DistancesFromArchitecture.hpp"
-#include "TokenSwappingWithArch/NeighboursFromArchitecture.hpp"
+#include "Architecture/DistancesFromArchitecture.hpp"
+#include "Architecture/NeighboursFromArchitecture.hpp"
 
 using std::vector;
