+++ tket/tests/TokenSwapping/TestUtils/PartialTsaTesting.hpp
@@ -15,7 +15,7 @@
 #pragma once
 
 #include "TokenSwapping/PartialTsaInterface.hpp"
-#include "TokenSwappingWithArch/ArchitectureMapping.hpp"
+#include "Architecture/ArchitectureMapping.hpp"
 #include "Utils/RNG.hpp"
 
 namespace tket {
